lmrob.c: if SETUP_EQUILIBRATION fails, do not call error() anymore.
dgeequ fails for matrices that contain zero columns or rows that are
exaclty zero. Matrices with zero rows are sometimes created during the
calculation of M-S estimates.

The behaviour in case of zero rows has been changed to give a warning only
and then simply skip the equilibration step.

Added an illustration of the problem in tests/m-s-estimator.R.

Added a comment in ?lmrob.M.S.

git-svn-id: svn://svn.r-forge.r-project.org/svnroot/robustbase@365 694ef91d-65df-0310-b7bb-92e67a308ead